Add support for GPU normal map

+++ packages/dev/core/types/dem/dem.tileclient.d.ts
@@ -3,8 +3,8 @@ import { FetchResult, ITileAddress, ITileClient, ITileMetrics } from "../tiles/t
 import { IDemInfos } from "./dem.interfaces";
 export declare class DemTileWebClient implements ITileClient<IDemInfos> {
     _elevationsDataSource: ITileClient<Float32Array>;
-    _normalsDataSource?: ITileClient<Float32Array>;
-    constructor(elevationSrc: ITileClient<Float32Array>, normalSrc?: ITileClient<Float32Array>);
+    _normalsDataSource?: ITileClient<Uint8ClampedArray>;
+    constructor(elevationSrc: ITileClient<Float32Array>, normalSrc?: ITileClient<Uint8ClampedArray>);
     get metrics(): ITileMetrics;
     fetchAsync(request: ITileAddress, ...userArgs: unknown[]): Promise<FetchResult<Nullable<IDemInfos>>>;
     private computeNormals;
